fixed validators and changed price to float

+++ backend/db/migrations/20231128160717-create-event.js
@@ -49,7 +49,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.NUMERIC,
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       startDate: {
